+++ instagram_to_discord/discord_event_listener.py
@@ -11,9 +11,8 @@ from .instagram_type import get_multiple_medias_from_str
 from .sites.tiktok_handler import handle_tiktok_main
 from .sites.youtube_handler import handle_youtube_main
 from .sites.instagram.instagram_process import process_instagram
-from .sites.twitter.twitter import send_twitter_images_from_cache_for_specified_index
+from .sites.twitter.twitter import send_twitter_images_from_cache_for_specified_index, twitter_line_to_image_urls
 from .sites.twitter.twitter_process import process_twitter
-from .twitter_multiple import twitter_line_to_image_urls
 from .util import is_int
 
 IS_DEBUG = os.getenv("IS_DEBUG")
