--- instagram_to_discord/twitter/__init__.py
@@ -1,2 +1,0 @@
-from .base64_util import base64_decode_str, base64_encode_str
-from .twitter_image import TwitterImage, convert_twitter
