+++ tests/test_twitter.py
@@ -6,7 +6,7 @@ import unittest
 from debug import DEBUG
 
 from instagram_to_discord.instagram_type import convert_long_caption
-from instagram_to_discord.twitter.twitter_image import convert_twitter
+from instagram_to_discord.sites.twitter.twitter_image import convert_twitter
 
 from . import context
